Finished bin reallocation and hash-space adaptation
Changed compspot back to its original file, and moved processing accounting outside, back in the strategy, modified strategy to include hash-space re-allocation method and changed the way in which requests are @
'handled'. Note that processing maximums and minimums are strictly based on each 'bin'/hash-space, rather than both repo CPU usage AND hash-space-based CPU usage. Next up: explain algorithm, sort put errors remaining and make sure that messages that should not be stored are not stored in repos/are routed towards where they are relevant (not depending on labels anymore, but rather h-spaces).