All working as intended - no cloud, no bucket duplicates
Changed quite a lot! Actually, now registering hits with cache checks, as well (in the case that the (processed) content is not in storage, but is in cache. Re-routing unprocessed data back to the responsible EDRs, until deadline expires, only AFTER sending them to the cloud. Also changed content placement for that. Request rate increased by a lot!