Working perfectly now, with functions instantiated and 'available' everywhere
Everything works well and to plan (now). Functions are now instantiated as VMs instantly, whenever the admit_task method is called, thus services are 'callable' everywhere that the needed 'hashes' are available in storage. NEXT: update all the parts of the no-reuse branch and then do the dataset workload tests